Require cocotb >= 1.5, < 2
Update the version constraint for cocotb to at least version 1.5, but
don't explicitly allow 2.x versions yet until we know that they are
compatible.

cocotb 1.5 was released on 2021-03-11 and is backwards-compatible with
older 1.x releases, so users shouldn't have much trouble updating to
this version to continue using the next released version of
cocotb-coverage.